+++ src/components/AuthRoute/AuthRoute.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 
 import { Navigate, useLocation } from 'react-router-dom';
 
-import { useAppSelector } from '../../services/hooks';
+import { useAppSelector } from '../../hooks/hooks';
 
 interface AuthRouteProps {
 	children: ReactNode;
