+++ src/components/ProtectedRoute/ProtectedRoute.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
-import { useAppSelector } from '../../services/hooks';
+import { useAppSelector } from '../../hooks/hooks';
 
 interface ProtectedRouteProps {
 	children: ReactNode;
