+++ src/hooks/hooks.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from './store';
+import type { RootState, AppDispatch } from '../services/store';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
