+++ src/pages/Login/Login.tsx
@@ -11,7 +11,7 @@ import { InputProps } from '../../types';
 
 import { loginUser, clearError } from '../../services/slices/authSlice';
 
-import { useAppDispatch, useAppSelector } from '../../services/hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks/hooks';
 
 import styles from './styles.module.css';
