+++ src/pages/Profile/Profile.tsx
@@ -15,7 +15,7 @@ import {
 	getUserProfile,
 } from '../../services/slices/authSlice';
 
-import { useAppDispatch, useAppSelector } from '../../services/hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks/hooks';
 
 import styles from './styles.module.css';
